media: Set null CcLayer when entering PictureInPicture mode
Currently when a video enters picture-in-picture (PIP) mode, the
SurfaceLayer is still compositing in the original window, in addition to
the PIP window. Typically this is not visible to users because it is
fully occluded by the <div> created by the PictureInPictureInterstitial
which is opaque. But in corner cases (e.g. during window resizing), the
video could still be composited, which could cause scaling issues when
the video is using DCompTexture on Windows.

This CL sets a null CcLayer when the video enters PIP mode, and
sets back the original CcLayer when exiting PIP mode to solve this
issue.

Manually tested and the scaling bug goes away. Also tried entering and
exiting the full screen mode, normal (inline) mode and PIP mode and
the video is playing as expected.

Bug: 1385649
Test: See above
Change-Id: I9d96416a99c98661ccca58a044a12f47f737a570
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087511
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083146}
NOKEYCHECK=True
GitOrigin-RevId: 54bc384e36b9b4055807e8df74664f8e644cb79d