Fix fill-mode-missing-from-to-keyframes
The main issue with this test was that we are checking styles on a
running animation. Though the test had some allowance for slack it
was not nearly sufficient.  There is absolutely no need for slack if
we instead start the animation in a paused state.

A second issue with the test is that it was trying to test too much
and actually tested less than the name suggested.  The test was
checking combinations of fill-mode, missing from and/or to keyframes,
and play-direction. That results in 48 combinations.  The test was
creating 10 animations, but only checking 5 of them. With the patch,
we drop consideration of the play direction and exhaustively test the
remaining combinations.  This choice aligns with the name of the test.

Bug: 1350341
Change-Id: I7e012c2b7f669856d2b6fa9d04ad2091e714d538
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4105267
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083193}
NOKEYCHECK=True
GitOrigin-RevId: f17626485c7a3b3c3848361d90413006503bb248