bluetooth: Remove expectation for filter-does-not-match.https.html
A timeout expectation was added for this test in the
virtual/web-bluetooth-new-permissions-backend test suite but not for the
normal version. This provided an opportunity to observe that while the
virtual test stopped flaking on 2022-11-17 when the expectation change
landed the non-virtual test stopped flaking on 2022-11-21.

Neither test is now flaking so this expectation can be removed.

Fixed: 1385858
Change-Id: Ie8cdd8378d1d60a9786e5240b2c9a3febd0e8184
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108051
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Commit-Queue: Alvin Ji <alvinji@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083268}
NOKEYCHECK=True
GitOrigin-RevId: 711e0856d22e481797193a7e6a425ff4d8d93f32