Accumulate relative insets correctly in OOF fragmentation.
We are dealing with accumulated relative *insets*, not offsets, so they
shouldn't include containing block offsets. Just swap top, left,
inline-start, block-start values and flip signs as needed. That's all we
should do.

This was mainly an issue if we had mixed 'direction' (LTR vs RTL). With
mixed writing-modes, stuff becomes monolithic, which disengages the
fragmentation machinery and instead performs regular out-of-flow layout.
I am wondering if it in theory could also be a problem for mixed
writing-modes, if a writing-mode root isn't the containing block of an
OOF descendant, but I couldn't come up with a such a testcase that fails
without this fix.

Bug: 1395581
Change-Id: Id572d33440b08743da2e3bc2a3758c1ed5f4be56
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4106468
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083339}
NOKEYCHECK=True
GitOrigin-RevId: a39fde6225fda2f19481edc88b08734b7de6322e