Remove data race causing OffscreenCanvas placeholder to freeze
In this previous CL, new logic was introduced to make the release of
placeholder frames happen when the last refcount is released:
https://chromium-review.googlesource.com/c/chromium/src/+/3826420
In that CL it was possible for resources to be recycled directly in
CanvasResourceRasterSharedImage::OnBitmapImageDestroyed, which was
causing the resource to be recycled without going through
CanvasResourceDispatcher::ReclaimResource, in the rare cases where
this callback was called after the next frame has been received by
the placeholder canvas. This is because the bitmap was holding the
last reference to the canvas resource and was passing it instead of
releasing it. Therefore, the mechanism that returns resources after
last unref was not kicking in. This race condition resulted in an
off-by-one error (many of which can accumulate over time) in the
accounting of the number of unreturned placeholder frames, causing
CanvasResourceDispatcher to stop sending frames to the placeholder
canvas.

BUG=1393703

Change-Id: I20dbfb78d610266b7bbd3f38d638f7a6b423af5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4062751
Reviewed-by: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083378}
NOKEYCHECK=True
GitOrigin-RevId: 3f5ca566acdd1b18f306441d66373fd6f6e8cdb6