Make gradient use SkColor4f instead of SkColor
There has been a big refactor of changing instances of SkColor in favor
of SkColor4f, in order to better support CSS Color 4 and CSS Color 5,
including wide gamut devices.

We forgot an important instance in gradients that stores internally
SkColor.

The css-color4 test was broken. The color(srgb x x x) range is from 0
to 1, not from 0 to 255. This was getting a color way out of gamut and
trying to create a gradient.

More tests for gradients will be added in follow up CLs.

This CL changes that and all necessary other places in order to do the
change.

Bug: 1393174
Change-Id: Ib27203d3ca2ab936871e2abfeb794bc95e00db2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4053726
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Auto-Submit: Juanmi Huertas <juanmihd@chromium.org>
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083404}
NOKEYCHECK=True
GitOrigin-RevId: a7d60235d014c955aa8c096f011d6f7aa6e6c083