CSS: Unbox InterpolableColor.
Currently InterpolableColor holds 4 unique_ptrs to doubles and a
unique_ptr to a vector of 4 unique_ptrs to doubles. This results in a
lot of allocations. In fact, according the a pprof profile collected
on Speedometer2, this is the top contributor to the number of cycles
spent in malloc(). The CL simply unboxes all the leaf allocations.

Speedometer2.Total improvement:
M1: -0.6%
Linux: -0.7%

Bug: 1400639
Change-Id: I88f974e9111b3a7bf1a6e828aaee02c604ae358d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4100966
Commit-Queue: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1083481}
NOKEYCHECK=True
GitOrigin-RevId: d6fb0dfe837f0e2719ad37c6b9d1d27b3b3ef2e5