[BREFs] Use modified_runtime_features in RuntimeFeatureStateOverrideContext
Initialize the RuntimeFeatureStateOverrideContext with the
modified_runtime_features navigation parameter.

The modified_runtime_features is sent from the browser and contains
all the feature override values that the override context needs to
expose to the renderer.

These values will be applied to the override context very soon after
the ExecutionContext is constructed.

Bug: 1377000
Change-Id: I52219bf23b4a33b5b85978329f0b89a4b7e36f59
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4089492
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084012}
NOKEYCHECK=True
GitOrigin-RevId: 9aac9a8ee8a5e41eaed19ed2400c0c7fb6b09c8f