Add console message for resources with IDNA deviation chars in hostnames
Chrome is planning to switch from IDNA 2008 Transitional Mode to
Non-Transitional Mode. This change will result in some hostnames
resolving to different IP addresses if the hostname contains an IDNA
deviation character. See
https://chromestatus.com/feature/5105856067141632 for details.

For example, a page containing a subresource like
<img src="http://faß.de/image.jpg"> will no longer resolve to
http://fass.de/image.jpg in Non-Transitional Mode. We want to warn
web developers in advance if they embed a resource with an affected
URL.

This CL adds a console message on pages embedding such resources if
the requestor origin is different than the resource origin (so that
we don't warn on an affected domain itself).

This is non-trivial because URL processing code presently doesn't keep deviation characters intact. As such, they aren't present in KURLs.
This means we can't simply check the hostname of the KURL for
deviation characters.

This CL adds an additional field to KURL to mark it if the original URL
string as present in the HTML of the page contains deviation characters.
It then decodes the hostname as IDNA if necessary, before printing the
log message. These URLs should be extremely rare, so this shouldn't
introduce any significant overhead; memory or performance wise.

All of this code is temporary and should be removed once
Non-Transitional IDNA is enabled in Chrome.

Bug: 694157
Change-Id: I339934bd143194e7763ec326389fde6aa764b555
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080821
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084016}
NOKEYCHECK=True
GitOrigin-RevId: ee1a5bd8a900dcc56c2f3e29e224634c43f6e5ac