Split image-set rendering tests.
Currently the WPT image-set rendering test contains several sub-tests.

Because it is a reference test, it generates a pass or fail of
the overall file, needing all sub-tests to pass for the test to pass.

This does not provide any granularity into which parts of the spec are
currently failing to render and which are passing.

Splitting the rendering tests into individual tests, aiming to have a
single failure reason for each test, should provide a better signal on
how far along each implementation is and make it easier to track
progress towards full spec implementation.

The image-set-type-unsupported-rendering tests were fixed to conform
to the spec for unsupported types.

Spec definition:
"An image-set() function contains a list of one or more
<image-set-option>s, and must select only one of them to determine what
image it will represent:
  First, remove any <image-set-option>s from the list that specify
  an unknown or unsupported MIME type in their type() value."

If all the values in the image set are of an unsupported type, the set
should be empty.

Spec link:
https://w3c.github.io/csswg-drafts/css-images-4/#image-set-notation

R=masonf, nrosenthal, pdr

Change-Id: Ic862112d5f8eb963620c840c3e49c3916c7e2d73
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4103915
Reviewed-by: Philip Rogers <pdr@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Traian Captan <tcaptan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084025}
NOKEYCHECK=True
GitOrigin-RevId: 87da37160ce98a019c47f8eb44f91ec49b750236