Revamp the popover "nearest open ancestral popover" concept
This CL simplifies the logic for "nearest open ancestral popover" and
reduces the scope to just the showPopover use case. The documentation
is improved (fixing crbug.com/1349611) and simplified. For light
dismiss, it turns out the big algorithm wasn't needed - a simple
set of DOM ancestor checks will do the trick. To accomplish the
simplified ancestor algorithm, an HTMLCollection is added to Document
for all popover invokers.

Two of the more convoluted tests for ancestor relationships were
updated with this logic, since they now work slightly differently.
One is that a popover nested *inside* an invoking element no longer
connects that nested popover to the invoked popover. The other is
primarily a fix to the test, which was intending to ensure an "unused"
invoker still participates in the ancestor tree.

Bug: 1307772
Fixed: 1349611
Change-Id: I68c1223acd7a4aa1b6781162d747db04deab1d65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4099790
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084044}
NOKEYCHECK=True
GitOrigin-RevId: 1e52a850fcee317ab5e2b963cf035bed6945c490