pip2: only reuse player if it is picture-in-picture related.
Currently, we are reusing player for any cases if two documents have the
same origin when PIP2 is enabled. This could cause security issue
because they don't have lifetime guarantee therefore the player could
outlive the page.

We will only reuse player if either the old document is in pip window
or the new document is in pip window. There will be a follow up CL that
guarantee these two documents have opener/pip relationship. And because
the opener window's lifetime strictly contains the pip window's, we are
safe to move the player between them.

Bug: 1392441
Change-Id: Ide3824de509ebf7d517b1f33dad69f1941764c6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4068769
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Jazz Xu <jazzhsu@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084092}
NOKEYCHECK=True
GitOrigin-RevId: 73a4c3a8300bc6552d1f95aad4c2a458bad83837