[IndexedDB] Add `IndexedDBClientStateChecker`
This chain of CLs aims to provide a way for the IndexedDB service to
query the state of its client without introducing addition dependencies
of browser domain concepts like `RenderFrameHost` or back/forward cache
to the IndexedDB service, which minimize the work for the future work
when the IndexedDB need to be taken out as a standalone service
running on a separate process.

An example usage of this checker is for IDB process to determine if
the IDB client is currently inactive and thus not supposed to receive
certain events. See https://crrev.com/c/3980739.

Design doc: https://docs.google.com/document/d/182KlafoJMGeYe48GkOTMuJ76WhXMNmOvlp79eWEYkiA

In summary, this CL mainly does the following changes:
1. Create a mojom interface `IndexedDBClientStateChecker` which is
used as the protocol between the IDB service and the browser process.
It also uses an empty `IndexedDBKeepActive` interface to scope the
active period of the client.

2. Create two implementations for the `IndexedDBClientStateChecker`
interface: `NoDocumentIndexedDBClientStateChecker`, which is a default
checker that is used when the client has no document associated, and
`DocumentIndexedDBClientStateChecker` which is attached to a document,
when the mojo method is invoked, it will call
`RenderFrameHost::IsInactiveAndDisallowActivation` to destroy the
document if it's inactive (except for prerendering).

Bug: 1331187
Change-Id: I1f4a09c23b324d41b7c442161e8f3212fb1d5cc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3980833
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Fergal Daly <fergal@chromium.org>
Commit-Queue: Mingyu Lei <leimy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084162}
NOKEYCHECK=True
GitOrigin-RevId: 1eec91b9300b5a2990b6d6e904b50e016e0f64b4