Fix DCHECK with Text*coderStream
When write() or close() was called on a TextEncoderStream or
TextDecoderStream in a detached realm it would cause a DCHECK failure
due to the context being invalid. Check for an invalid context and throw
an exception proactively.

BUG=1395588

Change-Id: Iafab2d90a2a35da01dfd811e4d0501a2971a5db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4080291
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Commit-Queue: Domenic Denicola <domenic@chromium.org>
Auto-Submit: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084191}
NOKEYCHECK=True
GitOrigin-RevId: 8c09c47c64aefe4ad282616112e519efec45ffca