[Storage Partitioning] Update virtual test suite to disable partitioning
This CL is the first of two to ultimately enable the
ThirdPartyStoragePartitioning feature whenever the
EnableExperimentalWebPlatformFeatures flag is specified.

This CL updates the existing third-party-storage-partitioning
virtual test suite such that storage partitioning is disabled
for the suite. This entailed copying the non-virtual test
expectation files into the virtual test suite folders. This
makes the virtual test suite redundant temporarily, but in the
next CL we will enable ThirdPartyStoragePartitioning whenever
EEWPF is enabled and move all the test-expectation files replaced
in this CL into the non-virtual test folders.

Bug: 1346609

Change-Id: I7514c93673a427683c201559437dba29bc483e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4094178
Commit-Queue: Andrew Williams <awillia@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084318}
NOKEYCHECK=True
GitOrigin-RevId: c86d8720c54aec8c62387aab0dbb6b6d9cf3f506