Speculative fix for scroll-timeline-default-iframe
Only flaking on a Mac, but the failure looks very similar to flakes
on other scroll-animation reftests, and believe the same fix will take
care of it.  The problem found in other tests was that the screenshot
was being taken too soon, while compositor commits were still being
blocked on FCP. This patch adds waits in both the iframe and parent
frame. The child frame waits for the compositor to be ready (and 2
extra frames) before telling the parent it is ready.  The parent
follows the same set of rules before activating the screenshot.

Bug: 1358133
Change-Id: I273be3358a6325e8512e8b782eefaae506f15f53
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108428
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084413}
NOKEYCHECK=True
GitOrigin-RevId: 4940e2d0365e23290315c7d44a1212ed62e0854b