[GridFragmentation] Store placement data upon HasBlockFragmentation
In https://chromium-review.googlesource.com/c/chromium/src/+/4109928,
we added support to store grid placement data on `NGGridLayoutData`.

This was originally stored when `HasBlockFragmentation`. During review,
this was moved to `InvolvedInBlockFragmentation`, but when
`IsBreakInside` is false. This was supposed to be `IsBreakInside`
is true. This didn't impact the test case or my original repro, so it
seemed safe. However, now that the change has landed, I retested the
live site and saw that it's still crashing.

Moving the data storage under the true case for `IsBreakInside`
fixes the live site.

Bug: 1399336
Change-Id: Ic702d2ae229ef23d582d75d98c0860be55e69ccb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112318
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1084467}
NOKEYCHECK=True
GitOrigin-RevId: 67934e0a229000984b8450752d58de405e1142dd