[CompositeClipPathAnimations] Fix behavior for custom timing functions
Previously, custom timing functions affected the float animation used
as input to the clip-path paint worklet. This caused a situation where the clip path paint worklet would be confused as to which keyframe the animation was on.

This change forces the float animation to use linear timing
functions, and applies the timing functions within the paint worklet.

Bug: 1374390

Change-Id: Idede83e0b75c84690a0b6a6d3cee4fc94a8771dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3975531
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084492}
NOKEYCHECK=True
GitOrigin-RevId: 459de11eb46ce1c6ae40131ed7f0c4e48ddb544f