Web Locks: Reject promise for non-fully-active documents
The script-facing request() and query() methods in our Web Locks
implementation had an early exit for the non-fully-active a.k.a.
detached document case, but simply returned a dummy promise. Since
this was implemented, the web platform generally and the Web Locks API
specifically have better defined the behavior for this case, where a
rejected promise should be returned.

There was already a WPT for this which we were failing. Make it pass!

BUG: 1234679

Change-Id: I6ad937292c2d768fba783b890f3bb4332cb228cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112406
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Joshua Bell <jsbell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084539}
NOKEYCHECK=True
GitOrigin-RevId: 49083fdb056bdedf355e4581dba930bf5237f084