Reland r1081902 with additional geometry mapper cache invalidation
This patch is a reland of https://crrev.com/1081902 ("Only invalidate
GeometryMapper cache if paint properties change") which caused several
bugs (crbug.com/1400638) and was reverted in https://crrev.com/1083413.
The root-cause of these bugs was a missing invalidation of the geometry
mapper cache when SetParent was called on isolation/alias nodes. The
geometry mapper cache stores ancestor information and needs to be
updated when ancestors change, even if they are isolation/alias nodes.
Prior to r1081902, this under invalidation was hidden because the
PrePaint tree walk always invalidated the geometry mapper cache.

Patchset 1 is the original patch, and the additional patchsets include
the changes made on top of that.

Change-Id: Icfe5f8e41f5dc064b716e6c4847075dfdefcb8bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4114275
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084579}
NOKEYCHECK=True
GitOrigin-RevId: 10b66992ceb72d40b97f342ef9dd5086ef6765d9