Added default resolution support in css image-set
Currently, if the resolution of an image-set value is missing, parsing
fails.

Based on the spec, the resolution is optional and should default to 1x
if not specified.

Spec definition:
"A <resolution> (optional). This is used to help the UA decide which
<image-set-option> to choose. If the image reference is for a raster
image, it also specifies the image’s natural resolution, overriding any
other source of data that might supply a natural resolution.

If not specified, it behaves as 1x for the purpose of selecting which
<image-set-option> to use. It also defaults the image’s natural
resolution to 1x, but if some other source of data supplies a natural
resolution, that resolution must be honored instead."

https://w3c.github.io/csswg-drafts/css-images-4/#image-set-notation

R=masonf, pdr

Fixed: 1400898
Change-Id: I6d15966d88307cfdfa6ce00e9fca3a64fcfceea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112418
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Traian Captan <tcaptan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084682}
NOKEYCHECK=True
GitOrigin-RevId: a69cb4c787fd4dfb1bb8b253869b3e849c780d91