Open-code SVGAnimatedProperty::CreateAnimatedValue()
This function is only called in one place - when creating the underlying
value for a SMIL animation - and all implementations do the same thing.
Moving this out of the property interface will make the animation code
less spread out. In the future the SVG OM types may not be used for SMIL
animations.

Use a switch on the property type in the animation code instead of this
virtual type-dispatching function.

Add a CreateUnderlyingValueForAnimation() to the SVGAnimateElement
hierarchy to hold this operation.

Bug: 1017723
Change-Id: Ibaa00953183c774519760f2a3950870bbcdd8b75
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4111947
Auto-Submit: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1084705}
NOKEYCHECK=True
GitOrigin-RevId: a48521701f582ceb2f92df65b135001ab9f55d03