Fenced frame: fix MPArch histogram resize unit test
This test never worked for MPArch. Doing a full navigation in a unit
test using the MPArch implementation causes the fenced frame to attempt
sending a mojo message to the associated remote fenced frame[1].
However, the browser-side remote is never set up, since this is just a
unit test. That causes a DCHECK() to fail[2], and the test crashes.

The reason we never hit this crash before the ShadowDOM removal is
because `HTMLFencedFrameElementTest` was never actually testing MPArch
before the ShadowDOM removal CL. While the class used
`testing::WithParamInterface` to get the implementation, it was never
actually used, and both the "MPArch" and ShadowDOM parameterized tests
had the implementation type hard-coded to `shadow_dom`[3].

This CL updates the test to not navigate to any page, and instead
manually sets the variable that tells the fenced frame to freeze the
frame size.

[1]https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/html/fenced_frame/fenced_frame_mparch_delegate.cc;l=39;drc=25a1416a6d94b0b26cb925a2c3be288ed361a219
[2]https://source.chromium.org/chromium/chromium/src/+/main:mojo/public/cpp/bindings/lib/interface_endpoint_client.cc;l=573;drc=38c289882d3328f77f23492b6383cbe0c5265dbd
[3]https://chromium-review.googlesource.com/c/chromium/src/+/3918971/51/third_party/blink/renderer/core/html/fenced_frame/html_fenced_frame_element_test.cc

Change-Id: I9cdfdd956bd35c5cdc46febb0070c62d423c40ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112539
Commit-Queue: Liam Brady <lbrady@google.com>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085003}
NOKEYCHECK=True
GitOrigin-RevId: 0d934e77029623d9f3639905f22a562e42d042c7