[Perf] Fixed perf regression introduced on CL:4083950
Bug: 1400871
The perf regression was introduced as a part of a bug fixed,
this CL fixes the regression by altering the order of some
`if` checks as well as for the return statement so that no
unnecessary expensive work is done for elements that do not
need it.

The latest Pinpoint job results for this CL can be found
in the CRBug comments.
The CL also adds a missing comment to the header file,
that should've been added on the original CL.

Change-Id: I42ddc66aeded982cd651d8d344124b12a24124dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4108989
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1085168}
NOKEYCHECK=True
GitOrigin-RevId: c695f018269f3d44c74cfd299cd0d42e8ed7ae87