Handle UnguessableToken::Deserialize failure in /third_party/blink/renderer/bindings/core/v8
We plan to change UnguessableToken::Deserialize so that it
returns an absl::optional<UnguessableToken> instead of an
UnguessableToken, with the optional not containing a value
in the case where an empty UnguessableToken would be
returned today. One side effect of this is that for empty
UnguessableTokens, attempting to access it without checking
if a value exists will result in a crash, whereas today an
empty UnguessableToken can be returned (when DCHECKs are not
enabled) and used for some things without realizing it.

This CL goes through non-test cases where we don't check
whether UnguessableToken::Deserialize returns an empty
token and attempts to add the check + appropriate handling. In
all cases I tried to have the code fail gracefully, but if
some of these can truly never cause
UnguessableToken::Deserialize to create an empty
UnguessableToken, we could replace this logic with a CHECK.

Once this CL lands, the next step will be to change the return
type of UnguessableToken::Deserialize and update all usages.

This CL was uploaded by git cl split.

R=japhet@chromium.org

Bug: 1402549
Change-Id: I6dc860c18bdcd6cad1b891efb949e86a08879f8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4116599
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Auto-Submit: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085214}
NOKEYCHECK=True
GitOrigin-RevId: 3f3b2d1b051cdaeffd16eb993065cd81f9489f9c