Drop 16x16 alignment requirements for AndroidVideoEncodeAccelerator.
Similar to the change for NdkVideoEncodeAccelerator. The only
difference is the Java MediaCodec has a reset() method so we don't
have to completely destroy and recreate the MediaCodec to change
to an aligned size.

With this change, the alignment modifier in RTCVideoEncoder can also
be removed.

R=eugene, ilnik

Bug: 719023, 1084702, 1383643
Change-Id: Ida24ca8b872e86f2667160f432bf84cdb4116397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4112291
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085266}
NOKEYCHECK=True
GitOrigin-RevId: 2f9bccf66e35b6fba1c881433ffdc5a8059dc6f3