RTCVideoEncoder: Executes UpdateEncoderInfo() in media sequence
If UpdateEncoderInfo() is done before the first GetEncoderInfo()
after InitEncode(), UpdateSetRates() tries to change a hardware
encoder L1T1. But this is denied by the hardware encoder because
the temporal layers change in spatial layer encoding is not
supported. This mismatches the metadata of bitstream buffer and
cause the software fallback.
This CL avoids the problem by making UpdateEncoderInfo() run
the media sequence to update EncoderInfo immediately.

BUG=b:261437029
TEST=webrtc.RTCPeerConnectionPerf.vp9_svc_l2t3_270p_hw

Change-Id: I4acf1b7165e6f519db3b17756bc5b98f95215cb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4087968
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Henrik Boström <hbos@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085310}
NOKEYCHECK=True
GitOrigin-RevId: 14789944a2e09b5728f725ce439d54d11d9b3f95