Implement software codec selection for NdkVideoEncodeAccelerator.
This adds a formal "is_software_codec" flag for VideoEncodeAccelerator
implementations to report support for. Clients must set
required_encoder_type on the Config to kSoftware or kNoPreference in
order to actually be given a software codec.

The new modifier will be shown on chrome://gpu for debugging
purposes. The Android and macOS VEA implementations will also
emit a MEDIA_LOG indicating a software codec has been selected.

Notes:
- Bitrate modes aren't accurate on software codecs. They're just given
a default of VBR in MediaCodecInfo if nothing is specified in the OS
level codec XML, but seem to happily output CBR.
- The software encoders appear to only support constrained profiles,
which we were failing to enumerate in CodecProfileLevelList.
- Since NdkVideoEncodeAccelerator is enforcing hardware codec selection
now, the tests and RTC video encoder have to be updated to set
kNoPreference now.

Bug: 719023, 1084702, 1383643
Change-Id: I71dce1524c9b727f31e57e4bc1ac0b538abbf853
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4083807
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085325}
NOKEYCHECK=True
GitOrigin-RevId: a86d7aae276e24789f9b54ce3827d1c363055526