currentcolor support for color-mix
color-mix values are generally resolved into single blink::Colors at
computed value time. "currentcolor" cannot do this as its value is
unknown until after style is computed.

In the event that a color-mix contains "currentcolor" this CL adds
a UnresolvedColorMix struct to StyleColor that will get
resolved at used value time when currentcolor is available. The
internal color in a StyleColor can now either be a blink::Color
or a pointer to an UnresolvedColorMix.

If neither color is a "currentcolor" (or a color-mix value with a
"currentcolor") we continue to resolve color-mix values at computed value time.

It is still unresolved what we'll do when the value of both percentages
for color-mix are zero as it should get rejected at parse time but
there are no WPT tests that cover this situation yet.

Changes to third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-computed-color-mix-function-expected.txt are due
to new tests passing. New failures are all old failures just the
message has changed.

The approach uses union type for blink::Color and a unique pointer to
an UnresolvedColorMix together. This means that there is a lot more
boilerplate code and potentially some difficult-to-maintain custom
memory management. A version of this CL that uses absl::variant is
here crrev.com/c/4091814

Running the size command on content_shell for linux for the various
versions of this CL:
Main: 146058981
Here (union and unique ptr): 146077125 (delta: 18,144)
crrev.com/c/4091814: 146110917 (delta: 51,936)

Bug: 1333988
Change-Id: I954f2371bc267e7912b62dbaccc363e5bdf901da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4017892
Commit-Queue: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085475}
NOKEYCHECK=True
GitOrigin-RevId: 208e2b1e0578ad51ff148171027047c4b6b771a3