[FedCM] Add login hint behind a flag
This CL adds a working prototype of the loginHint parameter in the
IdentityProviderConfig dictionary. This behavior is gated by a flag that
is disabled by default. The parameter allows the caller to request only
accounts matching the given loginHint in either the account ID or email
to be shown.

There is still an open question on how loginHint should behave with the
IDP Signin Status API, but both are experimental features at the moment.

Bug: 1356021
Change-Id: I5b18d2a662624007f1ca79ed409b6f39929ac782
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4104301
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Yi Gu <yigu@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085543}
NOKEYCHECK=True
GitOrigin-RevId: 8e55856ab640966448dd9c8e016cd5130c54ece4