Plumbing initiator base url into NavigationRequest.
This is the first of a series of CLs to implement using the base
url of the initiator for about:srcdoc and about:blank frames.
See further details at
https://docs.google.com/document/d/1e7T1YR5aGDg-eGHKDNnKUWcz1Dr38t_O0-XJqsMeZcE/edit?usp=sharing&resourcekey=0-qCAYJPulnTdo9hV_dPCdhw

This CL implements snapshotting the initiator's base url (for frame
load requests) at the same time the initiator origin is captured,
and then plumbs it back to the start of NavigationRequest via
CommonParams. The plumbing in this CL is not guaranteed to be
exhaustive, so there may be additional plumbing added in the follow-on
CLs.

This CL has no behavioural changes.

Change-Id: I37759aba16d47e12300b6f11f3f6374ad3362261
Bug: 1356658, 751329
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4026883
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: James Maclean <wjmaclean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085791}
NOKEYCHECK=True
GitOrigin-RevId: 23e90a11b9094fe2682083445f14abbaea9bef48