[StyleBuilder] Lazily compute IsStackingContextWithoutContainment
IsStackingContextWithoutContainment is a field that is derived from
other information on ComputedStyle. In theory, we could perform the
calculation in full on every access, but since there are lots of
checks involved, we store the result as a flag on ComputedStyle.
This also simplifies some diff functions, since we'd otherwise
need to track all field which this flags depend on (now and in the
future).

However, the current approach of updating this flag during
Element::StyleForLayoutObject is not compatible with the StyleBuilder
project, since we're done building the style by then. Creating a new
StyleBuilder just for this flag is also not desirable, since the flag
is commonly set for situations where performance is very relevant
(e.g. many elements with transforms/opacity).

This CL instead proposes a new field type "derived_flag", which
allows lazily computing a flag based on ComputedStyle, and caching
the result.

Derived flags can be diffed like normal fields (as seen from the
perspective of computed_style_diff_functions.json5). Because derived
flags are required to be top-level (i.e. exist directly on
ComputedStyle), we do not need to enumerate and maintain a huge list
of field_dependencies: the comparison will always be run by diffing
functions that list IsStackingContextWithoutContainment as a field.

Bug: 1377295
Change-Id: I250bda001aafcba549cfb6e72822331d9c730f82
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4110776
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1085850}
NOKEYCHECK=True
GitOrigin-RevId: 2d43c5a30bf49014981c061e955abe44e1e43117