Move ownership of DataSource into DemuxerManager
The current implementation still has WMPI controlling the destruction
of the data source, but in future CLs this will be changed so that
DemuxerManager will be the only thing posted for destruction by WMPI,
and manage the destruction of DataSource and Demuxer internally.

Additionally, the helper that gets a raw pointer to DataSource will be
removed in the change set when the demuxers will be owned by
DemuxerManager directly.

Bug: 1377053
Change-Id: I5830cca5d9d1c8905a97d83e187246d3d19fbbfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3999802
Reviewed-by: Matthew Wolenetz <wolenetz@chromium.org>
Commit-Queue: Ted (Chromium) Meyer <tmathmeyer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1086198}
NOKEYCHECK=True
GitOrigin-RevId: 7154f630fd2a312c55153f305f50b58ae01f0c92