Revert tentative change and add more dcheck
This CL reverts part of a tentative fix:
https://chromium-review.googlesource.com/c/chromium/src/+/4009789
That part was trying to create a copy of the texture before sending
it to the async part. This didn't fix the clusterfuzz issue so we are
reverting it. The rest of the tentative fix was adding DCHECK, we are
leaving those as can be helpful to validate future issues.

And also add a new DCHECK validating that image is properly working.

Bug: 1381881
Change-Id: Iab28439d907109b6eef73fb86e5bcdb5f53f9a5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4117630
Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1086475}
NOKEYCHECK=True
GitOrigin-RevId: 4e9af64907ae9b1c011339c9c674d4db488902ba