[DevTools] Remove bogus SASS sourcemap test.
This test is trying to ensure that sourcemaps can never be stale, but
does so in a way that doesn't match what happens in reality (adding
revisions to a UISourceCode) and peeks into the internals of the
SourceMapManager. The latter makes it unnecessary to challenging to
actually fix the sourcemap staleness issue (which the test doesn't
prevent anyways), so we'll remove this test here and instead rely on
the appropriate tests on the DevTools side.

Bug: chromium:508270, chromium:1403362
Change-Id: Ia70a347a3eeda17c2c6c9348bc14ca8fa7069762
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4120506
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Quick-Run: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1086672}
NOKEYCHECK=True
GitOrigin-RevId: a05e348d4d4501d70ee69bcd794956e2fdd03942