Only disable further fragmentation if fragmentation is on.
With the new call to DisableFurtherFragmentation() in
NGOutOfFlowLayoutPart::GenerateFragment(), we might attempt to disable
fragmentation even if we weren't participating in block fragmentation (a
monolithic OOF, for instance). This would lead to a crash if the
constraint space didn't have a rare-data object.

Bug: 1403467
Change-Id: Icd2e005440f93263511ff68949d40672d4030dd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4127783
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1088065}
NOKEYCHECK=True
GitOrigin-RevId: b06d93bfa21444888d6c31505de49aceb06efb96